+++ source/package.json
@@ -14,6 +14,7 @@
     "react-redux": "^5.1.1",
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.1",
+    "reactstrap": "^6.5.0",
     "redux": "^4.0.1",
     "redux-action": "^1.2.2",
     "redux-pender": "^2.0.9"
