+++ source/src/components/Top.jsx
@@ -1,5 +1,7 @@
 import React from 'react';
-import styles from '../styles/Top.scss';
+import styles from 'styles/Top.scss';
+
+//import Stylesheet
 import classNames from 'classnames'
 
 const cx = classNames.bind(styles);
