--- source/src/components/common/AppContents.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import AppProfile from './AppProfile';
-import PostList from './PostList';
-
-const AppContents = () => {
-    return (
-        <div>
-            <AppProfile/>
-            <PostList/>
-        </div>
-    );
-};
-
-export default AppContents;
