+++ source/src/components/common/ContentsTemplate.jsx
@@ -1,12 +1,17 @@
 import React from 'react';
-import ProfileContainer from './ProfileContainer';
-import AppContainer from './AppContainer';
+import ProfileContainer from './ProfileContainer/ProfileContainer';
+import AppContainer from './AppContainer/AppContainer';
+
 
 const ContentsTemplate = () => {
   return (
     <div>
+      <div>
       <ProfileContainer/>
+      </div>
+      <div>
       <AppContainer/>
+      </div>
     </div>
   );
 };
