--- source/src/components/common/ProfileContainer.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import ProfileTop from './ProfileTop';
-import ProfileBottom from './ProfileBottom';
-
-const ProfileContainer = () => {
-  return (
-    <div>
-      <ProfileTop/>
-      <ProfileBottom/>
-    </div>
-  );
-};
-
-export default ProfileContainer;
