+++ source/src/styles/Base.scss
@@ -1,5 +1,5 @@
 body {
-  background-color: #e0e0e0;
+  background-color: hsl(0, 0, 80%);
   margin: 0;
 }
